fix: 연도별 금융기관 지원금액 평균값의 소수점 제거

+++ src/main/java/com/ehem/kakaopay/model/guarantee/service/GuaranteeService.java
@@ -113,7 +113,7 @@ public class GuaranteeService {
         List<AverageAmountPerYearResult> amountResults =
                 guaranteeRepository.findAverageAmountsPerYear().stream()
                         .filter(averageAmountPerYearResult -> averageAmountPerYearResult.getInstitute().getName().equals(instituteName))
-                        .sorted(Comparator.comparingDouble(AverageAmountPerYearResult::getAvg))
+                        .sorted(Comparator.comparingInt(AverageAmountPerYearResult::getAvg))
                         .collect(Collectors.toList());
 
         AverageAmountPerYearResult minResult = amountResults.get(0);
